+++ src/client/Fenetre5.java
@@ -38,7 +38,7 @@ String s1 ="";
 			 String ouest = "west";
 			 String est = "east";
 
-			 int emp = f.getEmp();
+			 long emp = f.getEmp();
 			 int car = f.getCarinthetown();
 			 Double tp= f.getTp();
 			 Double td=f.getTd();
